+++ fstrack-tractor-api/src/users/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './update-first-time.dto';